Refactor models

+++ models/orm_models/income.py
@@ -3,7 +3,7 @@ from app import database
 
 class Income(database.Model):
     id = database.Column(database.Integer, primary_key=True)
-    sum = database.Column(database.Integer(150), unique=False)
+    sum = database.Column(database.Integer, unique=False)
     type = database.Column(database.String(150), unique=False)
     creation_date = database.Column(database.DateTime)
